producer/requirements.txt
certifi==2025.4.26
charset-normalizer==3.4.2
holidays>=0.50
idna==3.10
kafka-python==2.2.7
numpy==2.2.5
pandas==2.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.3
six==1.17.0
tzdata==2025.2
urllib3==2.4.0
